Platform width and two sides to the game

+++ Lift/core/src/com/lift/game/controller/entities/EntityBody.java
@@ -33,7 +33,7 @@ public class EntityBody {
      * @param body
      * @param sensor
      */
-	protected final void add_fixture(Body body, float[] vertexes, int width, int height, float density, float friction, float restitution, short category, short mask, boolean sensor) {
+	protected final void add_fixture(Body body, float[] vertexes, float width, float height, float density, float friction, float restitution, short category, short mask, boolean sensor) {
         for (int i = 0; i < vertexes.length; i++) {
             if (i % 2 == 0) vertexes[i] -= width / 2;
             if (i % 2 != 0) vertexes[i] -= height / 2;
